python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.9.3",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
]
